+++ apt-pkg/rpm/rpmrecords.cc
@@ -438,7 +438,7 @@ void rpmRecordParser::GetRec(const char *&Start,const char *&Stop)
    headerGetEntry(HeaderP, RPMTAG_ARCH, &type, (void **)&str, &count);
    BufCatTag("\nArchitecture: ", str);
 
-   snprintf(buf, sizeof(buf), "%lu", Handler->FileSize());
+   snprintf(buf, sizeof(buf), "%llu", (unsigned long long) Handler->FileSize());
    BufCatTag("\nSize: ", buf);
 
    BufCatTag("\nMD5Sum: ", Handler->MD5Sum().c_str());
